Refactoring the common part of TTreeReaderGenerator and TTreeProxyGenerator.
